# cmd/api/docs/swagger.yaml
definitions:
  models.OTPData:
    properties:
      phone:
        type: string
    required:
    - phone
    type: object
  models.Product:
    properties:
      category_id:
        type: integer
      colour:
        type: string
      id:
        type: integer
      price:
        type: string
      product_image:
        items:
          type: integer
        type: array
      product_name:
        type: string
      stock:
        type: string
    type: object
  models.UserDetails:
    properties:
      confirmpassword:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        default: false
        type: boolean
    type: object
  models.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.UserSignInResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      role:
        default: false
        type: boolean
      user_id:
        type: integer
    type: object
  models.VerifyData:
    properties:
      code:
        type: string
      phone:
        type: string
    required:
    - code
    - phone
    type: object
info:
  contact: {}
paths:
  /admin/userlist:
    post:
      consumes:
      - application/json
      description: Retrive and display user list
      produces:
      - application/json
      responses:
        "200":
          description: 'Array of user details '
          schema:
            items:
              $ref: '#/definitions/models.UserDetails'
            type: array
        "400":
          description: Bad request
          schema:
            items:
              $ref: '#/definitions/models.UserDetails'
            type: array
      security:
      - ApiKeyHeaderAuth: []
      summary: ListProducts
      tags:
      - User Management
  /user/otplogin:
    post:
      consumes:
      - application/json
      description: verify Phone number using OTP
      parameters:
      - description: Phone number
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.OTPData'
      produces:
      - application/json
      responses:
        "200":
          description: phone number
          schema:
            items:
              $ref: '#/definitions/models.UserSignInResponse'
            type: array
        "400":
          description: Bad request
          schema:
            items:
              $ref: '#/definitions/models.UserSignInResponse'
            type: array
      summary: SendOTP
      tags:
      - User Auth
  /user/signin:
    post:
      consumes:
      - application/json
      description: Sign in a user and return user details and a token
      parameters:
      - description: User login details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: User details and role
          schema:
            $ref: '#/definitions/models.UserSignInResponse'
        "400":
          description: Bad request
          schema:
            items:
              $ref: '#/definitions/models.UserSignInResponse'
            type: array
      summary: UserSignIN
      tags:
      - User Auth
  /user/verifyotp:
    post:
      consumes:
      - application/json
      description: verify Phone number using OTP
      parameters:
      - description: Phone number and code
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.VerifyData'
      produces:
      - application/json
      responses:
        "200":
          description: phone number
          schema:
            items:
              $ref: '#/definitions/models.UserSignInResponse'
            type: array
        "400":
          description: Bad request
          schema:
            items:
              $ref: '#/definitions/models.UserSignInResponse'
            type: array
      summary: VerifyOTP
      tags:
      - User Auth
  /user/viewproducts:
    post:
      consumes:
      - application/json
      description: view products by a user
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad request
          schema:
            items:
              $ref: '#/definitions/models.Product'
            type: array
      summary: ViewProducts
  /users/signup:
    get:
      consumes:
      - application/json
      description: Retrive UserDetails stored in DB and a auth token with success
        message
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserDetails'
      produces:
      - application/json
      responses:
        "200":
          description: User details and token
          schema:
            items:
              $ref: '#/definitions/models.UserDetails'
            type: array
        "400":
          description: Bad request
          schema:
            items:
              $ref: '#/definitions/models.UserSignInResponse'
            type: array
      summary: UserSignUP
      tags:
      - User Auth
swagger: "2.0"